src/test/java/guru/springframework/converters/UnitOfMeasureCommandToUnitOfMeasureTest.java
package guru.springframework.converters;

import guru.springframework.commands.UnitOfMeasureCommand;
import guru.springframework.domain.UnitOfMeasure;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.*;

public class UnitOfMeasureCommandToUnitOfMeasureTest {

    private final UnitOfMeasureCommandToUnitOfMeasure converter = new UnitOfMeasureCommandToUnitOfMeasure();

    @Before
    public void setUp() throws Exception {

    }

    @Test
    public void testNull() {
        assertNull(converter.convert(null));
    }
    @Test
    public void testEmptyObject() {
        assertNotNull(converter.convert(new UnitOfMeasureCommand()));
    }

    @Test
    public void convert() {
        //given
        UnitOfMeasureCommand uomCommand = new UnitOfMeasureCommand();
        uomCommand.setDescription("testDesc");
        uomCommand.setId(1L);
        //when
        UnitOfMeasure uom = converter.convert(uomCommand);
        //then
        assertNotNull(uom);
        assertEquals("testDesc",uom.getDescription());
        assertEquals(1, (long) uom.getId());
    }
}